Cast some floats to ints in magnet_tools.py
Currently, if mpi_runner.py is run on a cluster then no errors occur,
but if it's done on some other machine then the use of floats in array
slicing causes errors.

This difference is because when mpi_runner.py is run on a cluster it's
done through mpijob.sh, and one of the things that script does is (along
with "module load python/ana") it sources the mpi2 environment which
loads numpy version 1.9.0. However, doing just "module load python/ana"
by itself loads numpy version 1.14.2.

It seems that the two different versions of numpy handle floats as array
indices differently, but really there shouldn't be any floats as array
indices in the first place.