Add --seed and --seed_value options to mpi_runner.py
Allow the option of seeding the random number generator used in the
algorithm with a specific number. This is useful because it allows the
ability to have the same output genomes if the same seed_value is chosen
for multiple executions of mpi_runner.py, and this will be quite useful
for running tests on the genome files outputted.