Test creation of genome files from shimming
Move the process of creating the genome files for a shim to its own
function to enable testing of it.

Add parameter to saveh5() function for the lookup variable (previously
the lookup variable was a global variable but is now defined inside the
process() function, so it needs to be explicitly passed to saveh5()
now).

In ID_Shim_BCell.generate_children(), the equality check of real_bfield
with None threw an error if real_bfield was a numpy array, so use
"is None" instead of "== None" for the check.

Add test data for shimming.