Remove old imports related to dls-python
The python version that was originally used to run OptID was the
Controls Group's dls-python which made use of require() to import
dependencies, but at some point there was a switch to using python from
anaconda and so the dls-python imports via require() are no longer
needed.

Also, tidy up the imports in the mpi_runner_for_shim_opt.py script and
remove some unused imports.