Update tests to reflect sort data reorganisation
Due to mpi_runner.py with the -r flag taking a directory to look for
genome files in, previously test_process_initial_population() was
accidentally pulling in
1.07788212e-08_000_c0833a96b82c.genome.inp.genome as one of the genome
files to start off the population, but this is effectively just a copy
of 1.07788212e-08_000_c0833a96b82c.genome so it'd be better to leave
that file out and only use the genome files created from the output of
the sort process. The simplest way to achieve this was to move the
genome files outputted by the sort process to their own directory, which
is the new mpi_runner_output/ directory.

The location change of the genome files required some of the other tests
to be changed slightly to reflect their new location.