Sort genome filenames before loading initial population
Previously, when travis ran the test
MpiRunnerTest.test_process_initial_population() it failed despite the
test passing on a local machine. This is possibly due to the different
ordering of listed directory contents output between the travis job and
the local test run when getting the genome filenames to append to the
"population" list, causing "population[0]" to be different between the
travis job and the local test run:
https://github.com/DiamondLightSource/Opt-ID/blob/e52d6cf8c1d8be0ec775786c65f9db8b29085f7c/IDSort/src/mpi_runner.py#L143

Try sorting the list of genome filenames so then "population[0]" will
consistently be the same filename, given a directory with the same
contents on different operating systems.