Update magnets.py to python 3
For a first attempt at updating magnets.py to python 3, just get the
magnets_test.py test to pass when using python 3.7.

One change in particular is due to the change from cPickle to pickle:
the pickled object outputted by cPickle.dump() is in general different
from the pickled object outputted by pickle.dump(). Even though there
are some possible sidesteps around this to make pickled objects
compatible between python 2 and 3, as there is currently no intention of
supporting both python 2 and 3 once the conversion to python 3 has been
done it seemed more sensible to just regenerate the test_cpmu.mag file
in the test data repo to be the python 3 equivalent of what it was
before.