Fix a few deprecation warnings when running tests
In h5py, using the syntax "dataset.value" has been deprecated in favour
of "dataset[()]", so just a change to the new sytax.

In numpy, linalg.lstsq() warns that the default value of the "rcond"
parameter will change in the future and advises to explictly use the
future default value to check if your code will be broken when that
change is implemented. Using the future default of "rcond=None" doesn't
break OptID so that change has been made.

Travis also gave a warning about the use of h5py.File() without
explicitly passing what mode the file should be opened in, so the file
mode has been added in the relevant places.