Add output dir parameter to process_genome.py
Instead of having the output files of process_genome.py be written to
the current directory, pass it a path to a specific directory.

Also, make minor changes to some variable names in process_genome.py to
make it a bit clearer when filepaths are being handled as opposed to
filenames.