Generate script for an existing-population sort job
There are some functionalities provided by OptID that don't follow the
entire process of a sort or shim but are still useful to users, such as
running a sort job but starting with an existing genome population
instead of creating an initial population from scratch. In that
particular case it's not necessary to generate the json, mag, h5/lookup
files as they should already exist from the creation of the initial
population and so only one particular script is needed to be executed to
perform this task.

However, one of the aims of introducing a config file was to alleviate
users from having to interact with individual scripts in OptID and to
just have one central script that is used for everything.

One way around having users run that individual script in OptID and
making them need to keep track of what parameters were used for the
initial sort to then pass again for this other job was to autogenerate a
shell script after a sort job that is aware of all the relevant
parameter values and can just be executed without any extra information
to do the existing-population sort job.