Write cluster job logfiles to a fixed directory
Without any configuration there are multiple logfiles from the job
submitted to GridEngine, as well as the logfiles being written to
whatever the current working directory OptID is run in is.

These changes configure the multiple logfiles to be merged into one
file, and they are written to the logfiles/ directory which is a fixed
child directory of the chosen data directory passed to OptID. The
logfile names are also now just the JOB_ID of the associated job that
generated them instead of "the script name passed to qsub plus the
JOB_ID".