Enable naming of genome report files
Genome report files are now written to the genome_reports/ directory
inside the main data directory. The generate_report.sh script may now be
passed the --report-filename flag to specify the filename of the PDF
genome report. Alternatively, when this flag is omitted the report will
be named by concatenating all filenames passed to generate_report.sh
with an underscore character as the separator.