Autogenerate mpijob.sh and mpi4shimOpt.sh scripts
As there are no plans for the moment to create a modulefile/install for
Opt-ID, the code discovery of Opt-ID is done on a checkout-by-checkout
basis. Currently the mpijob.sh and mpi4shimOpt.sh scripts have hardcoded
paths to a particular checkout of the Opt-ID repo in the filesystem, but
they need to instead have the path to wherever the cloned repo is in the
filesystem.

Template these scripts to then pass in the correct Opt-ID checkout path.
However, rather than having the users manually generate the mpijob.sh
and mpi4shimOpt.sh scripts, autogenerate them when the relevant job is
run.