Use only one .inp file to start a shim job
Previously, in the YAML config of a shim, the
process_genome.readable_genome_file value was a list which implied that
it was possible to start a shim job from multiple .inp files. However,
this likely isn't a use case, so instead have its value be a single
string to reflect that shim jobs would really only be started from a
single .inp file.